Added processing for delete comments

+++ src/Controller/PostCommentController.php
@@ -7,6 +7,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Doctrine\Common\Collections\Collection;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 class PostCommentController extends AbstractController
 {
@@ -15,6 +17,12 @@ class PostCommentController extends AbstractController
      */
     public function delete(Request $request, PostComments $postComments): Response
     {
+
+        if($this->getUser() !== $postComments->getAuthor() && !in_array("ROLE_ADMIN",$this->getUser()->getRoles()))
+            return $this->redirectToRoute('post_show', [
+                'id' => $postComments->getPost()->getId(),
+            ]);
+
         if ($this->isCsrfTokenValid('delete'.$postComments->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($postComments);
